+++ scovid19/lib/data/__init__.py
@@ -1,0 +1,3 @@
+from scovid19.lib.data.Scotland import Scotland
+from scovid19.lib.data.Infections import Infections
+from scovid19.lib.data.Vaccines import Vaccines
